requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.7.0
Werkzeug==2.3.6
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.7.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
